pkg/*: Fix trivial unconvert and ineffassign issues.
See,
$ golangci-lint run --disable-all --deadline 10m --enable=goconst --enable=deadcode --enable=gofmt --enable=ineffassign --enable=structcheck --enable=unconvert --enable=varcheck
pkg/services/quota/quota.go:51:13: unnecessary conversion (unconvert)
				if int64(usedSessions) > scope.DefaultLimit {
				        ^
pkg/api/login.go:166:44: unnecessary conversion (unconvert)
	decryptedError, err := util.Decrypt([]byte(decoded), setting.SecretKey)
	                                          ^
pkg/services/auth/auth_token.go:152:2: ineffectual assignment to `needsRotation` (ineffassign)
	needsRotation := false
	^

+++ pkg/services/quota/quota.go
@@ -48,7 +48,7 @@ func (qs *QuotaService) QuotaReached(c *m.ReqContext, target string) (bool, erro
 					return false, err
 				}
 
-				if int64(usedSessions) > scope.DefaultLimit {
+				if usedSessions > scope.DefaultLimit {
 					c.Logger.Debug("Sessions limit reached", "active", usedSessions, "limit", scope.DefaultLimit)
 					return true, nil
 				}
